@@ -2,47 +2,34 @@
     "private": true,
     "scripts": {
         "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --disable-host-check --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "development": "mix",
+        "watch": "mix watch",
+        "watch-poll": "mix watch -- --watch-options-poll=1000",
+        "hot": "mix watch --hot",
         "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "eslint": "./node_modules/.bin/eslint resources/js/ test/ --ext .js"
+        "production": "mix --production",
+        "eslint": "./node_modules/.bin/eslint resources/js/ test/ --ext .js,.vue"
     },
     "devDependencies": {
-        "@babel/plugin-proposal-class-properties": "^7.10.4",
-        "@babel/preset-react": "^7.0.0",
-        "babel-plugin-react-scoped-css": "^1.1.1",
-        "bootstrap": "^4.0.0",
-        "cross-env": "^7.0",
-        "css-loader": "^5.2.6",
+        "@babel/preset-react": "^7.13.13",
         "eslint": "^7.28.0",
         "eslint-loader": "^4.0.2",
         "eslint-plugin-react": "^7.24.0",
-        "jquery": "^3.2",
-        "laravel-mix": "^5.0.1",
-        "laravel-mix-alias": "^1.0.3",
-        "lodash": "^4.17.19",
-        "node-sass": "^6.0.0",
-        "popper.js": "^1.12",
-        "resolve-url-loader": "^3.1.0",
-        "sass-loader": "^8.0.2",
-        "scoped-css-loader": "^1.0.0",
-        "style-loader": "^1.2.1",
-        "webpack": "^4.44.1"
+        "laravel-mix": "^6.0.6",
+        "webpack": "^5.38.1"
     },
     "dependencies": {
         "axios": "^0.21.1",
-        "dotenv": "^8.2.0",
+        "jquery": "^3.6.0",
         "prop-types": "^15.7.2",
         "react": "^16.14.0",
-        "react-date-picker": "^8.0.3",
+        "react-date-picker": "^8.2.0",
         "react-dom": "^16.14.0",
-        "react-naver-maps": "0.0.13",
-        "react-redux": "^7.2.0",
+        "react-is": "^17.0.2",
+        "react-naver-maps": "^0.0.13",
+        "react-redux": "^7.2.4",
         "react-router-dom": "^5.2.0",
-        "redux": "^4.0.5",
+        "redux": "^4.1.0",
         "redux-thunk": "^2.3.0",
         "styled-components": "^5.3.0",
         "styled-reset": "^4.3.4"
