@@ -1,6 +1,6 @@
 import React from "react";
 import styled, {css} from "styled-components";
-import color from "@/constant/color";
+import PropTypes from "prop-types";
 
 const styles = css`
     margin: 0;
@@ -11,5 +11,17 @@ const styles = css`
 const Heading = styled(({
     level, children, ...props
 }) => React.createElement(`h${level}`, props || null, children))`${styles}`;
+
+Heading.defaultProps = {
+    level: 4,
+    children: ''
+};
+
+Heading.propTypes = {
+    level: PropTypes.oneOfType([
+        PropTypes.number.isRequired,
+        PropTypes.oneOf([1, 2, 3, 4, 5, 6])
+    ])
+};
 
 export default Heading;
