@@ -1,6 +1,5 @@
 import React, {memo} from "react";
 import styled, {css} from "styled-components";
-import color from "@/constant/color";
 
 const styles = css`
     margin-top: 30px;
@@ -12,5 +11,10 @@ const Ul = styled.ul`${styles}`;
 const List = memo(({ordered, children, ...props}) => {
     return React.createElement(ordered ? Ol : Ul, props || null, children);
 });
+
+List.defaultProps = {
+    ordered: false,
+    children: ''
+};
 
 export default List;
