@@ -7,14 +7,15 @@ const StyledSpan = styled(Span)`
     display: block;
     position: absolute;
     top: -18px;
-    left: -18px;
+    left: -12px;
     width: 85px;
     height: 35px;
     border-radius: 6px;
+    background: ${color.pageColor};
     font-weight: bold;
     font-size: 12px;
     line-height: 35px;
-    color: ${color.blue600};
+    color: ${color.white};
     text-align: center;
     box-sizing: border-box;
 `;
@@ -26,5 +27,9 @@ const RideAttendInfo = memo(({children, ...props}) => {
         </StyledSpan>
     );
 });
+
+RideAttendInfo.defaultProps = {
+    children: ''
+};
 
 export default RideAttendInfo;
