@@ -1,23 +1,43 @@
 import React, {memo} from "react";
-import styled from "styled-components";
+import PropTypes from "prop-types";
+import {Link} from "react-router-dom";
+import styled, {css} from "styled-components";
 import LabelInput from "@components/UI/molecules/LabelInput";
 import Input from "@components/UI/atoms/Input";
 import color from "@/constant/color";
+import font from "@/constant/font";
 
 const ButtonWrapper = styled.div`
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
     overflow: hidden;
     margin-top: 20px;
     line-height: 40px;
 `;
 
+const ButtonStyles = css`
+    width: auto;
+    margin: 0 6px;
+    padding: 0 20px;
+    border-radius: 4px;
+    line-height: 40px;
+    font-size: ${font.sizeSmall};
+    text-decoration: none;
+    cursor: pointer;
+`;
+
 const StyledSubmitInput = styled(Input)`
-    display: block;
-    margin: 0 auto;
+    ${ButtonStyles};
     border: 0;
     background: ${color.pageColor};
-    line-height: 40px;
     color: ${color.white};
-    cursor: pointer;
+`;
+const StyledLink = styled(Link)`
+    ${ButtonStyles};
+    border: 1px solid ${color.pageColor};
+    background: ${color.white};
+    color: ${color.pageColor};
 `;
 
 const LoginForm = memo(({onSubmit, email, setEmail, password, setPassword}) => {
@@ -32,7 +52,7 @@ const LoginForm = memo(({onSubmit, email, setEmail, password, setPassword}) => {
                                  name: 'email',
                                  defaultValue: email,
                                  placeholder: '이메일을 입력해주세요.',
-                                 onChange: setEmail
+                                 onChange: event => setEmail(event.target.value)
                              }}/>
 
             <LabelInput isRequired={true}
@@ -44,15 +64,24 @@ const LoginForm = memo(({onSubmit, email, setEmail, password, setPassword}) => {
                                  name: 'password',
                                  defaultValue: password,
                                  placeholder: '비밀번호를 입력해주세요.',
-                                 onChange: setPassword
+                                 onChange: event => setPassword(event.target.value)
                              }}/>
 
             <ButtonWrapper>
                 <StyledSubmitInput type="submit"
                                    value="로그인"/>
+                <StyledLink to="/register">회원가입</StyledLink>
             </ButtonWrapper>
         </form>
     );
 });
+
+LoginForm.propTypes = {
+    onSubmit: PropTypes.func.isRequired,
+    email: PropTypes.string.isRequired,
+    setEmail: PropTypes.func.isRequired,
+    password: PropTypes.string.isRequired,
+    setPassword: PropTypes.func.isRequired
+};
 
 export default LoginForm;
