@@ -1,25 +1,25 @@
 export function formatDifficulty(difficulty) {
     switch (difficulty) {
         case 'beginner':
-            return '초보자'
+            return '초보자';
         case 'intermediate':
-            return '중급자'
+            return '중급자';
         case 'advanced':
-            return '숙련자'
+            return '숙련자';
         default:
-            return ''
+            return '';
     }
 }
 
 export function formatAltitude(altitude) {
     switch (altitude) {
         case 'flat':
-            return '평지'
+            return '평지';
         case 'uphill':
-            return '업힐'
+            return '업힐';
         case 'mountain':
-            return '산'
+            return '산';
         default:
-            return ''
+            return '';
     }
 }
