@@ -2,23 +2,5 @@
 
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
+Route::view('{path}', 'app')->where('path', '.*');
 
-Auth::routes([
-    'verify' => true,
-    'reset' => false
-]);
-
-Route::get('{path?}', [
-    'uses' => 'HomeController@index',
-    'where' => ['path' => '.*']
-]);
